moved image upload function into an action creator. sending location from aws s3 to store in itemreducer
