+++ app/containers/home/homeContainer.js
@@ -1,19 +1,21 @@
 import { connect } from 'react-redux';
 import { Home } from '../../components/home/home';
-import { itemsFetchData } from '../../actions/items/items';
+import { itemsFetchData, upload } from '../../actions/items/items';
 import { logoutUser } from '../../actions/session/actions';
 
 const mapStateToProps = state => ({
 	routes: state.routes,
 	user: state.sessionReducer.user,
-	items: state.itemReducer,
+	items: state.itemReducer.items,
+	recentUpload: state.itemReducer.recentUpload,
 	hasErrored: state.itemsHasErrored,
-	isLoading: state.itemsIsLoading,
+	isLoading: state.itemsIsLoading
 });
 
 const mapDispatchToProps =  {
   logout: logoutUser,
-  fetchData: itemsFetchData
+  fetchData: itemsFetchData,
+  upload
 };
 
 export default connect(mapStateToProps, mapDispatchToProps)(Home);
